requirements.txt
pillow>=11.2.1
pytesseract>=0.3.13
python-dotenv>=1.1.0
numpy>=2.0.2
pandas>=2.2.3